Introduce basic Rubinius support
This one didn't see the testing it deserved and can be implemented
better by using Rubinius internals, rather then this shaky
`binding_of_caller` frame guessing. For now, I wanna roll this out as a
experimental Rubinius support. We'll make it better post `2.0`.

I'm seeing gem installation errors on Travis at the moment. Both for
Rubinius and JRuby. I really wanna get them off the allowed failures, so
I'm gonna keep an eye out in the upcoming days.
